refactor: handle news in custom hook & update related files
