+++ src/app/[locale]/news/page.tsx
@@ -3,90 +3,14 @@
 'use client';
 
 import { FeaturedNewsCarousel } from '@/components/common/news/featured-news-carousel';
-import { useTranslations } from 'next-intl';
-import NewsCardProps from '@/components/common/news-card/types/news-card';
+import { useNews } from '@/lib/utils';
 
 /**
  * Highlighted news & general news
  * @returns {JSX.Element} - The news page
  */
 export default function News() {
-  const t = useTranslations('articleExcerpts');
-
-  // As far as I'm aware, this is the best way to do this. See https://next-intl.dev/docs/usage/messages#arrays-of-messages
-  const featuredNews: NewsCardProps[] = [
-    {
-      title: t('farmlinkPartnership.title'),
-      excerpt: t('farmlinkPartnership.excerpt'),
-      date: t('farmlinkPartnership.date'),
-      slug: t('farmlinkPartnership.slug'),
-      source: t('farmlinkPartnership.source'),
-      image: {
-        url: t('farmlinkPartnership.image.url'),
-        alt: t('farmlinkPartnership.image.alt'),
-      },
-    },
-    {
-      title: t('ambrookPartnership.title'),
-      excerpt: t('ambrookPartnership.excerpt'),
-      date: t('ambrookPartnership.date'),
-      slug: t('ambrookPartnership.slug'),
-      source: t('ambrookPartnership.source'),
-      image: {
-        url: t('ambrookPartnership.image.url'),
-        alt: t('ambrookPartnership.image.alt'),
-      },
-    },
-    {
-      title: t('regenerativeSeed.title'),
-      excerpt: t('regenerativeSeed.excerpt'),
-      date: t('regenerativeSeed.date'),
-      slug: t('regenerativeSeed.slug'),
-      source: t('regenerativeSeed.source'),
-      image: {
-        url: t('regenerativeSeed.image.url'),
-        alt: t('regenerativeSeed.image.alt'),
-      },
-    },
-  ];
-
-  const allNews: NewsCardProps[] = [
-    {
-      title: t('whyRegenerative.title'),
-      excerpt: t('whyRegenerative.excerpt'),
-      date: t('whyRegenerative.date'),
-      slug: t('whyRegenerative.slug'),
-      source: t('whyRegenerative.source'),
-      image: {
-        url: t('whyRegenerative.image.url'),
-        alt: t('whyRegenerative.image.alt'),
-      },
-    },
-    {
-      title: t('umichPartnership.title'),
-      excerpt: t('umichPartnership.excerpt'),
-      date: t('umichPartnership.date'),
-      slug: t('umichPartnership.slug'),
-      source: t('umichPartnership.source'),
-      image: {
-        url: t('umichPartnership.image.url'),
-        alt: t('umichPartnership.image.alt'),
-      },
-    },
-    {
-      title: t('ftcPartnership.title'),
-      excerpt: t('ftcPartnership.excerpt'),
-      date: t('ftcPartnership.date'),
-      slug: t('ftcPartnership.slug'),
-      source: t('ftcPartnership.source'),
-      image: {
-        url: t('ftcPartnership.image.url'),
-        alt: t('ftcPartnership.image.alt'),
-      },
-    },
-  ];
-
-  allNews.push(...featuredNews);
+  const { featuredNews } = useNews();
 
   return (
     <>
