--- src/data/featured-news.json
@@ -1,36 +1,0 @@
-{
-  "newsArticles": [
-    {
-      "id": "1",
-      "image": { "url": "/images/placeholder.jpg", "alt": "" },
-      "source": "Journal",
-      "date": "Apr 15, 2025",
-      "headline": "Todd Announces Partnership with Agricultural Innovation Lab",
-      "link": "news/todd-announces-partnership-with-agricultural-innovation-lab"
-    },
-    {
-      "id": "2",
-      "image": { "url": "/images/placeholder.jpg", "alt": "" },
-      "source": "PR Newswire",
-      "date": "Mar 30, 2025",
-      "headline": "Todd Introduces Revolutionary Regenerative Farming Techniques",
-      "link": "news/todd-introduces-revolutionary-regenerative-farming-techniques"
-    },
-    {
-      "id": "3",
-      "image": { "url": "/images/placeholder.jpg", "alt": "" },
-      "source": "AgTech Weekly",
-      "date": "Feb 12, 2025",
-      "headline": "Sustainable Agriculture Solutions Show Promising Results",
-      "link": "news/sustainable-agriculture-solutions-show-promising-results"
-    },
-    {
-      "id": "4",
-      "image": { "url": "/images/placeholder.jpg", "alt": "" },
-      "source": "Farm Journal",
-      "date": "Jan 28, 2025",
-      "headline": "Next Generation of Farming Technology Unveiled",
-      "link": "news/next-generation-of-farming-technology-unveiled"
-    }
-  ]
-}
